requirements.txt
discord.py==2.6.1
aiohttp==3.8.5
